+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-3500] Sign configuration secrets
+
 ## [4.19.0] 2021-05-04
 
 ### Changed